Try to do some speedup/cleanup
One thing this attempts to do is to -- Actually, it's bad. It was trying to load
from a file using the same object, but that is going to cause problems this way
so it will need to be reimplemented. The other speedup (and problem) was to use
the same quadtree instead of having to rebuild it. That is also wrong. Poop.